refactor(zbc): use intptr_t/uintptr_t for guest ints instead of intmax_t/uintmax_t
Changed function signatures in zbc_host.h (zbc_host_{read,write}_guest_int),
zbc_protocol.h (zbc_{read,write}_native_{int,uint}, zbc_parsed_t.parms/result),
and src/zbc_host.c implementations/casts for consistency with pointer-sized
values in guest/host interactions (e.g., addresses). Improves portability
across architectures where intmax_t exceeds pointer size.

+++ include/zbc_host.h
@@ -106,7 +106,7 @@ int zbc_host_process(zbc_host_state_t *state, uintptr_t riff_addr);
  * @param size  Size of integer in bytes
  * @return Integer value
  */
-intmax_t zbc_host_read_guest_int(const zbc_host_state_t *state,
+intptr_t zbc_host_read_guest_int(const zbc_host_state_t *state,
                                  const uint8_t *data, size_t size);
 
 /**
@@ -120,7 +120,7 @@ intmax_t zbc_host_read_guest_int(const zbc_host_state_t *state,
  * @param size  Size of integer in bytes
  */
 void zbc_host_write_guest_int(const zbc_host_state_t *state,
-                              uint8_t *data, uintmax_t value, size_t size);
+                              uint8_t *data, uintptr_t value, size_t size);
 
 #ifdef __cplusplus
 }
